Move tests, and exclude from final package
