+++ src/__tests__/fetch/index.test.ts
@@ -1,4 +1,4 @@
-import { tsFetch } from ".";
+import { tsFetch } from "../../fetch";
 
 // Mock the fetch API
 global.fetch = jest.fn();
