+++ src/__tests__/utility/helpers.test.ts
@@ -3,7 +3,7 @@ import {
   serializeQueryParams,
   QueryParams,
   isValidURL,
-} from "./helpers";
+} from "../../utility/helpers";
 
 describe("validateQueryParams", () => {
   it("should throw an error if params is not a plain object", () => {
@@ -41,8 +41,8 @@ describe("validateQueryParams", () => {
 describe("serializeQueryParams", () => {
   beforeEach(() => {
     // Mock the module containing validateQueryParams and ensure we can mock the function
-    jest.mock("./helpers", () => ({
-      ...jest.requireActual("./helpers"), // Retain all other functions as they are
+    jest.mock("../../utility/helpers", () => ({
+      ...jest.requireActual("../../utility/helpers"), // Retain all other functions as they are
       validateQueryParams: jest.fn(),
     }));
   });
